Refactoring API into separate package for easier management.
