+++ controllers/api/response.go
@@ -1,0 +1,22 @@
+package api
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+
+	log "github.com/gophish/gophish/logger"
+)
+
+// JSONResponse attempts to set the status code, c, and marshal the given interface, d, into a response that
+// is written to the given ResponseWriter.
+func JSONResponse(w http.ResponseWriter, d interface{}, c int) {
+	dj, err := json.MarshalIndent(d, "", "  ")
+	if err != nil {
+		http.Error(w, "Error creating JSON response", http.StatusInternalServerError)
+		log.Error(err)
+	}
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(c)
+	fmt.Fprintf(w, "%s", dj)
+}
