+++ controllers/phish.go
@@ -13,6 +13,7 @@ import (
 	"github.com/NYTimes/gziphandler"
 	"github.com/gophish/gophish/config"
 	ctx "github.com/gophish/gophish/context"
+	"github.com/gophish/gophish/controllers/api"
 	log "github.com/gophish/gophish/logger"
 	"github.com/gophish/gophish/models"
 	"github.com/gophish/gophish/util"
@@ -299,7 +300,7 @@ func (ps *PhishingServer) TransparencyHandler(w http.ResponseWriter, r *http.Req
 		SendDate:       rs.SendDate,
 		ContactAddress: ps.contactAddress,
 	}
-	JSONResponse(w, tr, http.StatusOK)
+	api.JSONResponse(w, tr, http.StatusOK)
 }
 
 // setupContext handles some of the administrative work around receiving a new
